+++ app/src/androidTest/java/com/example/StressOverflow/ItemListAdapterInstrumentedTest.java
@@ -8,12 +8,13 @@ import android.content.Context;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.platform.app.InstrumentationRegistry;
 
+import com.example.StressOverflow.Item.Item;
+import com.example.StressOverflow.Item.ItemListAdapter;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import java.util.ArrayList;
-import java.util.GregorianCalendar;
-import java.util.UUID;
 
 @RunWith(AndroidJUnit4.class)
 public class ItemListAdapterInstrumentedTest {
