+++ app/src/main/java/com/example/StressOverflow/AppGlobals.java
@@ -1,5 +1,7 @@
 package com.example.StressOverflow;
 
+import com.example.StressOverflow.Tag.Tag;
+
 import java.util.ArrayList;
