+++ app/src/main/java/com/example/StressOverflow/FilterDialog.java
@@ -2,9 +2,6 @@ package com.example.StressOverflow;
 
 import android.app.DatePickerDialog;
 import android.app.Dialog;
-import android.app.ListActivity;
-import android.text.Editable;
-import android.text.TextWatcher;
 import android.view.KeyEvent;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,20 +10,20 @@ import android.widget.AutoCompleteTextView;
 import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.CompoundButton;
-import android.widget.DatePicker;
 import android.widget.EditText;
 import android.widget.ListView;
 
+import com.example.StressOverflow.Item.Item;
+import com.example.StressOverflow.Item.ItemListAdapter;
+import com.example.StressOverflow.Tag.Tag;
 import com.google.android.material.button.MaterialButtonToggleGroup;
 import com.google.android.material.chip.Chip;
-import com.google.android.material.chip.ChipDrawable;
 import com.google.android.material.chip.ChipGroup;
 
 import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Objects;
