+++ app/src/main/java/com/example/StressOverflow/SignIn/SignUpActivity.java
@@ -1,6 +1,5 @@
-package com.example.StressOverflow;
-
-import static androidx.constraintlayout.helper.widget.MotionEffect.TAG;
+package com.example.StressOverflow.SignIn;
+import com.example.StressOverflow.Item.ListActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -15,16 +14,14 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.StressOverflow.R;
 import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.HashMap;
 import java.util.Map;
