+++ app/src/main/java/com/example/StressOverflow/Tag/AddTagFragment.java
@@ -1,4 +1,4 @@
-package com.example.StressOverflow;
+package com.example.StressOverflow.Tag;
 
 import android.app.Dialog;
 import android.content.Context;
@@ -7,13 +7,12 @@ import android.os.Bundle;
 
 import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.DialogFragment;
-import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
-import android.widget.EditText;
 import android.widget.TextView;
+
+import com.example.StressOverflow.R;
 
 /**
  * This fragment is called when a user wants to create a new tag
