+++ app/src/main/java/com/example/StressOverflow/Tag/AddTagToItemFragment.java
@@ -1,4 +1,4 @@
-package com.example.StressOverflow;
+package com.example.StressOverflow.Tag;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
@@ -7,26 +7,21 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 
-import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.Button;
 
+import com.example.StressOverflow.AppGlobals;
+import com.example.StressOverflow.Db;
+import com.example.StressOverflow.R;
 import com.google.android.material.chip.Chip;
 import com.google.android.material.chip.ChipGroup;
 import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.EventListener;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.FirebaseFirestoreException;
-import com.google.firebase.firestore.QueryDocumentSnapshot;
-import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 /**
  * Called when a user wants to add a tag to an already existing item
@@ -40,6 +35,7 @@ public class AddTagToItemFragment extends DialogFragment{
     private CollectionReference tagsRef;
     private Db tagDb;
     private String ownerName;
+
 
     public AddTagToItemFragment() {
         this.ownerName = AppGlobals.getInstance().getOwnerName();
