+++ app/src/main/res/layouts/Image/layout/select_image_fragment.xml
