+++ app/src/main/res/layouts/SignIn/layout/forgot_password_page.xml
