+++ app/src/main/res/layouts/Tag/layout/activity_tag_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TagList"
+    tools:context=".Tag.TagList"
     tools:visibility="visible">
 
     <ListView
