+++ app/src/main/res/layouts/Tag/layout/fragment_add_tag.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddTagFragment">
+    tools:context=".Tag.AddTagFragment">
 
 
     <EditText
