+++ app/src/main/res/layouts/Tag/layout/fragment_add_tag_to_item.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddTagToItemFragment">
+    tools:context=".Tag.AddTagToItemFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
